requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.9
django-environ==0.8.1
djangorestframework==3.12.4
idna==3.3
psycopg2-binary==2.9.2
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
prompt_toolkit==3.0.22
celery==5.1.2